requirements.txt
streamlit
google-generativeai
pandas
matplotlib